alpha_raspberry_pi_manager/requirements.txt
dotenv
python-telegram-bot
psutil
